pyproject.toml
[tool.poetry]
name = "sureflow"
version = "2.0.0"
description = "SureFlow — Batch KYC processing pipeline with Surepass API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }, { include = "server" }]

[tool.poetry.dependencies]
python        = "^3.10"
fastapi       = "^0.111"
uvicorn       = { version = "^0.30", extras = ["standard"] }
httpx         = "^0.27"
pandas        = "^2.2"
openpyxl      = "^3.1"
pyyaml        = "^6.0"
python-dotenv = "^1.0"
python-multipart = "^0.0.9"
psycopg = { version = "^3.2", extras = ["binary"] }

[tool.poetry.group.dev.dependencies]
pytest     = "^8.0"
pytest-cov = "^5.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
